Add support for multiple DEMs sharing SRTM 1|3s tiles
As we add more DEMs such as SRTM15s and Gebco, both which come with two flavors, there are many DEMS for which we need to switch to the same 1s or 3s tiles regardless of which dataset we requested.  This PR makes this change and makes it easier to add more later.